Add transparency support
I had to rely on a workaround in tc.py@export_writer:397 for rebuilding
the image, a temp image is saved and read to be able to retain alpha.

Apparently there's a bug in wxPython where the alpha channel is lost
when `GetSubBitmap` is used with `DrawBitmap`, there's a 'hack' where
calling `wx.PixelAlphaData` for that Bitmap works, but only if you don't
use masks, if masks are used the PixelData hack does not work.